For #20596 remove startup timeline probes
