+++ app/metrics.yaml
@@ -4463,129 +4463,6 @@ addons:
       - android-probes@mozilla.com
     expires: "2021-12-01"
 
-startup.timeline:
-  framework_primary:
-    disabled: true
-    send_in_pings:
-      - startup-timeline
-    type: timespan
-    time_unit: millisecond
-    description: |
-      The duration the Android framework takes to start before letting us run
-      code in `*Application.init` when this device has
-      `clock_ticks_per_second_v2` equal to 100: if it's not equal to 100, then
-      this value is captured in `framework_secondary`. We split this into two
-      metrics to make it easier to analyze in GLAM. We split on 100 because
-      when we did our initial brief analysis -
-      https://sql.telemetry.mozilla.org/queries/75591 - the results for clocks
-      ticks were overwhelmingly 100.
-
-      The duration is calculated from `appInitTimestamp -
-      processStartTimestamp`. `processStartTimestamp` is derived from the clock
-      tick time unit, which is expected to be less granular than nanoseconds.
-      Therefore, we convert and round our timestamps to clock ticks before
-      computing the difference and convert back to nanoseconds to report.
-
-      For debugging purposes, `clock_ticks_per_second_v2`, which may vary
-      between devices, is also reported as a metric
-    bugs:
-      - https://github.com/mozilla-mobile/fenix/issues/8803
-      - https://github.com/mozilla-mobile/fenix/issues/17972
-    data_reviews:
-      - https://github.com/mozilla-mobile/fenix/pull/9788#pullrequestreview-394228626
-      - https://github.com/mozilla-mobile/fenix/pull/15713#issuecomment-703972068
-      - https://github.com/mozilla-mobile/fenix/pull/18043#issue-575389284
-    data_sensitivity:
-      - technical
-    notification_emails:
-      - perf-android-fe@mozilla.com
-      - mcomella@mozilla.com
-    expires: "2021-08-01"
-  framework_secondary:
-    disabled: true
-    send_in_pings:
-      - startup-timeline
-    type: timespan
-    time_unit: millisecond
-    description: |
-      The duration the Android framework takes to start before letting us run
-      code in `*Application.init` when this device has
-      `clock_ticks_per_second_v2` not equal to 100. For more details on this
-      metric, see `framework_primary`
-    bugs:
-      - https://github.com/mozilla-mobile/fenix/issues/17972
-    data_reviews:
-      - https://github.com/mozilla-mobile/fenix/pull/18043#issue-575389284
-    data_sensitivity:
-      - technical
-    notification_emails:
-      - perf-android-fe@mozilla.com
-      - mcomella@mozilla.com
-    expires: "2021-08-01"
-  framework_start_error:
-    disabled: true
-    send_in_pings:
-      - startup-timeline
-    type: boolean
-    description: |
-      An error when attempting to record `framework_primary/secondary` - the
-      application init timestamp returned a negative value - which is likely
-      indicative of a bug in the implementation.
-    bugs:
-      - https://github.com/mozilla-mobile/fenix/issues/8803
-    data_reviews:
-      - https://github.com/mozilla-mobile/fenix/pull/9788#pullrequestreview-394228626
-      - https://github.com/mozilla-mobile/fenix/pull/15713#issuecomment-703972068
-    data_sensitivity:
-      - technical
-    notification_emails:
-      - perf-android-fe@mozilla.com
-      - mcomella@mozilla.com
-    expires: "2021-08-01"
-  framework_start_read_error:
-    disabled: true
-    send_in_pings:
-      - startup-timeline
-    type: boolean
-    description: |
-      An error when attempting to read stats from /proc pseudo-filesystem -
-      privacy managers can block access to reading these files -
-      the application will catch a file reading exception.
-    bugs:
-      - https://github.com/mozilla-mobile/fenix/issues/10434
-    data_reviews:
-      - https://github.com/mozilla-mobile/fenix/pull/10481
-      - https://github.com/mozilla-mobile/fenix/pull/15713#issuecomment-703972068
-    data_sensitivity:
-      - technical
-    notification_emails:
-      - perf-android-fe@mozilla.com
-      - mcomella@mozilla.com
-    expires: "2021-08-01"
-  clock_ticks_per_second_v2:
-    disabled: true
-    send_in_pings:
-      - startup-timeline
-    type: quantity
-    unit: clock ticks per second
-    description: |
-      The number of clock tick time units that occur in one second on this
-      particular device. This value is expected to be used in conjunction with
-      the `framework_primary/secondary` metrics.
-    bugs:
-      - https://github.com/mozilla-mobile/fenix/issues/8803
-      - https://github.com/mozilla-mobile/fenix/issues/18157
-    data_reviews:
-      - https://github.com/mozilla-mobile/fenix/pull/9788#pullrequestreview-394228626
-      - https://github.com/mozilla-mobile/fenix/pull/15713#issuecomment-703972068
-      - https://github.com/mozilla-mobile/fenix/pull/18158#issue-579593943
-    data_sensitivity:
-      - technical
-    notification_emails:
-      - perf-android-fe@mozilla.com
-      - mcomella@mozilla.com
-    expires: "2021-08-01"
-
 perf.startup:
   cold_main_app_to_first_frame:
     type: timing_distribution
