+++ app/src/main/java/org/mozilla/fenix/perf/StartupTimeline.kt
@@ -5,16 +5,6 @@
 package org.mozilla.fenix.perf
 
 import androidx.annotation.UiThread
-import androidx.lifecycle.Lifecycle
-import androidx.lifecycle.LifecycleObserver
-import androidx.lifecycle.OnLifecycleEvent
-import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.DelicateCoroutinesApi
-import kotlinx.coroutines.GlobalScope
-import kotlinx.coroutines.launch
-import mozilla.components.service.glean.private.NoReasonCodes
-import mozilla.components.service.glean.private.PingType
-import org.mozilla.fenix.GleanMetrics.Pings
 import org.mozilla.fenix.HomeActivity
 import org.mozilla.fenix.home.sessioncontrol.viewholders.topsites.TopSiteItemViewHolder
 import org.mozilla.fenix.perf.StartupTimeline.onApplicationInit
@@ -41,10 +31,7 @@ object StartupTimeline {
     private var state: StartupState = StartupState.Cold(StartupDestination.UNKNOWN)
 
     private val reportFullyDrawn by lazy { StartupReportFullyDrawn() }
-    internal val frameworkStartMeasurement by lazy { StartupFrameworkStartMeasurement() }
-    internal val homeActivityLifecycleObserver by lazy {
-        StartupHomeActivityLifecycleObserver(frameworkStartMeasurement)
-    }
+    internal val frameworkStartMeasurement by lazy { ApplicationInitTimeContainer() }
 
     fun onApplicationInit() {
         // This gets called from multiple processes: don't do anything expensive. See call site for details.
@@ -69,33 +56,5 @@ object StartupTimeline {
 
     private fun advanceState(startingActivity: StartupActivity) {
         state = StartupTimelineStateMachine.getNextState(state, startingActivity)
-    }
-}
-
-/**
- * A [LifecycleObserver] for [HomeActivity] focused on startup performance measurement.
- */
-@OptIn(DelicateCoroutinesApi::class) // GlobalScope usage
-internal class StartupHomeActivityLifecycleObserver(
-    private val frameworkStartMeasurement: StartupFrameworkStartMeasurement,
-    private val startupTimeline: PingType<NoReasonCodes> = Pings.startupTimeline,
-    private val scope: CoroutineScope = GlobalScope
-) : LifecycleObserver {
-
-    @OnLifecycleEvent(Lifecycle.Event.ON_STOP)
-    fun onStop() {
-        scope.launch { // use background thread due to expensive metrics.
-            // Ensure any last metrics are set before submission.
-            frameworkStartMeasurement.setExpensiveMetric()
-
-            // Startup metrics placed in the Activity should be re-recorded each time the Activity
-            // is started so we need to clear the ping lifetime by submitting once per each startup.
-            // It's less complex to add it here rather than the visual completeness task manager.
-            //
-            // N.B.: this submission location may need to be changed if we add metrics outside of the
-            // HomeActivity startup path (e.g. if the user goes directly to a separate activity and
-            // closes the app, they will never hit this) to appropriately adjust for the ping lifetimes.
-            startupTimeline.submit()
-        }
     }
 }
